858: Move profile visibility setting to profiles
