+++ lib/app/utils/profile_feature_checker.dart
@@ -9,7 +9,7 @@ import 'package:gruene_app/app/constants/design_constants.dart';
 import 'package:gruene_app/app/constants/secure_storage_keys.dart';
 import 'package:gruene_app/app/services/converters.dart';
 import 'package:gruene_app/app/services/gruene_api_profile_service.dart';
-import 'package:gruene_app/features/campaigns/screens/profile/profile_visibility_setting.dart';
+import 'package:gruene_app/features/profiles/widgets/profile_visibility_setting.dart';
 import 'package:gruene_app/i18n/translations.g.dart';
 import 'package:gruene_app/swagger_generated_code/gruene_api.swagger.dart';
 import 'package:http/http.dart';
